[refactor/#172] 애플 로그인 코드 VC -> VM 수정
