+++ Roomie/Roomie/Presentation/OnBoarding/ViewController/LoginViewController.swift
@@ -20,7 +20,7 @@ final class LoginViewController: BaseViewController {
     
     private let kakaoLoginButtonTapSubject = PassthroughSubject<Void, Never>()
     
-    private let appleLoginTokenSubject = PassthroughSubject<String, Never>()
+    private let appleLoginButtonTapSubject = PassthroughSubject<Void, Never>()
     
     private let viewModel: LoginViewModel
     
@@ -62,7 +62,7 @@ final class LoginViewController: BaseViewController {
             .tapPublisher
             .sink { [weak self] in
                 guard let self else { return }
-                self.performAppleLogin()
+                self.appleLoginButtonTapSubject.send()
             }
             .store(in: cancelBag)
     }
@@ -74,7 +74,7 @@ private extension LoginViewController {
     func bindViewModel() {
         let input = LoginViewModel.Input(
             kakaoLoginButtonTapSubject: kakaoLoginButtonTapSubject.eraseToAnyPublisher(),
-            appleLoginTokenSubject: appleLoginTokenSubject.eraseToAnyPublisher()
+            appleLoginButtonTapSubject: appleLoginButtonTapSubject.eraseToAnyPublisher()
         )
         
         let output = viewModel.transform(from: input, cancelBag: cancelBag)
@@ -88,53 +88,5 @@ private extension LoginViewController {
                 }
             }
             .store(in: cancelBag)
-    }
-    
-    func performAppleLogin() {
-        let appleIDProvider = ASAuthorizationAppleIDProvider()
-        let request = appleIDProvider.createRequest()
-        request.requestedScopes = [.fullName]
-        
-        let authorizationController = ASAuthorizationController(authorizationRequests: [request])
-        authorizationController.delegate = self
-        authorizationController.presentationContextProvider = self
-        authorizationController.performRequests()
-    }
-}
-
-// MARK: - ASAuthorizationControllerPresentationContextProviding
-
-extension LoginViewController: ASAuthorizationControllerPresentationContextProviding {
-    func presentationAnchor(
-        for controller: ASAuthorizationController
-    ) -> ASPresentationAnchor {
-        return self.view.window!
-    }
-}
-
-// MARK: - ASAuthorizationControllerDelegate
-
-extension LoginViewController: ASAuthorizationControllerDelegate {
-    func authorizationController(
-        controller: ASAuthorizationController,
-        didCompleteWithAuthorization authorization: ASAuthorization
-    ) {
-        guard let appleIDCredential = authorization.credential
-                as? ASAuthorizationAppleIDCredential else { return }
-        
-        let identityToken = appleIDCredential.identityToken.flatMap {
-            String(data: $0, encoding: .utf8)
-        }
-        
-        if let identityToken {
-            appleLoginTokenSubject.send(identityToken)
-        }
-    }
-    
-    func authorizationController(
-        controller: ASAuthorizationController,
-        didCompleteWithError error: Error
-    ) {
-        print("🚨Apple 로그인 실패: \(error.localizedDescription)")
     }
 }
